GameSync-AI/api/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
pyjwt
